node_modules: patchShebangs and symlink specified files during preinstall
This allows packages that ship scripts to execute their scripts during
the cooresponding install phase. It also allows us to link some files
realtive to the packages directory. Those are commonly required for
`bin-wrapper` packages that just wrap a binary, and if it isn't found
downloads them from the internet.