Remove path type assert on lock file
Various clean mechanisms return some attribute set with `outPath` and
not a `path` and even with `builtins.toPath` it will at best be a
"string with a context" but limiting the assert to a path does not seem
feasible. At the same time opening it being too many different things
seems not useful so it is probably best to just drop it.