Clarify usage of npmlock2nix
This helps beginners get started with npmlock2nix by clarifying the
README and fixing a few errors in it.

It adds a 'getting started' section that explains how to add npmlock2nix
to NIX_PATH, and replaces `nodelock2nix ? <FIXME>` with
`npmlock2nix ? import <npmlock2nix>` in the example `nix` files.