node_modules: support entries that have URLs as versions
In some cases a developer might pin a package to a specific URL in the
package.json. In that case the lockfile will not contain a resolved
key for the dependency but will have the original URL in the lockfile's
version field which we should use in that case.

With this patch I was able to build `Cinny` (a matrix client) on
revision 39b84a083d002deaa8f86689f97dbb887c27ffc0.