Patch the shebang of files specifed with bin field
Sometimes files specified with the bin field of package.json are missing
their executable bits when they are extracted. npm must be setting their
executable bits at some stage of `npm install` life cycle because by the
time the command is complete, they are executable. But at "preinstall"
stage, which is when we patch the shebangs, those files are not
executable thus skipped by the `patchShebangs` function.

We are now using `jq` to get the files specified with the "bin" field of
"package.json and patches the ones missing their executable bit.