[V2][bugfix] create missing dir in preInstallLinks script
Fixes the node-modules-attributes-are-passed-through unit test for the
lockfiles v2 implementation.

The directories only containing symlinked are getting deleted during
the tarball unpack process. In that particular test case, we symlink
cwebp to `./vendor/cwebp`. `./vendor` gets deleted after unpackaging,
meaning the build fails when trying to symlink nixpkgs's cwebp to
`./vendor/cwebp`: we need to re-create the vendor directory first.