[v2][bugfix] Remove https references from patched package.json
Some package dependencies can be expressed using a http link to a
tarball. Such as:

{
  "author": "",
  "dependencies": {
    "@matrix-org/olm": "https://gitlab.matrix.org/api/v4/projects/27/packages/npm/@matrix-org/olm/-/@matrix-org/olm-3.2.4.tgz"
  },
  "description": "",
  "devDependencies": {},
  "license": "ISC",
  "main": "index.js",
  "name": "url-as-version",
  "scripts": {
    "test": "echo \"Error: no test specified\" && exit 1"
  },
  "version": "1.0.0"
}

The version and other metadata attributes will be extracted by npm and
written in the package-lock.json file. For instance, for the previous
example:

{
  "dependencies": {},
  "lockfileVersion": 2,
  "name": "url-as-version",
  "packages": {
    "": {
      "dependencies": {
        "@matrix-org/olm": "3.2.4"
      },
      "license": "ISC",
      "name": "url-as-version",
      "version": "1.0.0"
    },
    "node_modules/@matrix-org/olm": {
      "integrity": "sha512-ddaXWILlm1U0Z9qpcZffJjBFZRpz/GxQ1n/Qth3xKvYRUbniuPOgftNTDaxkEC4h04uJG5Ls/OdI1YJUyfuRzQ==",
      "license": "Apache-2.0",
      "resolved": "file:///nix/store/qn1b7cpsw383kprpzvq4r1x3yis9bczn-olm-3.2.4.tgz",
      "version": "3.2.4"
    }
  },
  "requires": true,
  "version": "1.0.0"
}

The issue however, is that npm will invalidate the package-lock file
for each call and try to re-fetch the remote tarball to re-generate
it: there's no integrity mechanism in package.json, the remote URL
could have changed.

We go around that issue by replacing the URL in the version field of
package.json by a wildcard. Exactly like what we already do for github
dependencies.