Remove lockfilev1-specific machinery
The npm lockfiles v2 are meant to be retro-compatible with the "old"
v1 format. This retrocompatibility is implemented using using 2
top-level attributes: packages, the attribute containing the
v2-specific packages descriptions. dependencies, the attribute
containing the legacy packages descriptions.

{
   "packages": {
     "": { ... },
     "node_modules/left-pad": { ... }
   },
   dependencies: {
     ...
   },
   "version": ...
   ...
}

It turns out that npm won't complain if we drop entirely the
"dependencies" section as long as the description contained in the
"packages" attribute is complete-enough to build the node_modules
folder.

It's a good very news for us: it allows us to drop entirely support
for the v1 lockfile format in the v2 codebase. That's quite a massive
win on the long run: the machinery required to parse the v2 lockfile
is **much** simpler from the one required for the legacy format.

In this commit, we drop altogether all the support needed to parse the
lockfiles v1.

---------------------------

Loosely related: it turns out that npm disables the integrity check
when "null" is specified as an integrity. This behaviour is similar to
what we used to do except we don't have to calculate the store path
integrity anymore. It makes things a little bit easier code-wise.

We might have to revise this assumption if upstream end up changing
its behaviour in that regard.