[v2][bugfix] Patching shebangs of the bundled dependencies
Some npm dependencies are bundled, meaning their node_modules
directory is contained as part of their parent tgz distribution.

Like a regular package, a bundled package can contain some executable scripts.
That being said, these exec scripts won't be listed in the `bin`
section of their parent's package-lock.json file. This is to be
expected, those are transitive dependencies, they're not meant to be
listed there.

When patching a package through a source_override, we look whether or
not a node_modules directory is present. If it's there, it means we
potentially have some unpatched shebangs in these transitive
dependencies. We recurse in them using "patchShebangs".