[v2][bugfix] Relax package dependencies version bounds
Relaxing dependencies version bounds for github refs. If we happen to
leave them as is, it'll force NPM to checkout the remote repo to get
the actual ref.

At this point, we already pinned everything through the "resolved"
field, there's no point in keeping the unresolved reference in
dependencies. We can relax those to "*".