[v2][bugfix] Accept missing resolved link for bundled dependencies
Some bundled dependencies can end up not having a "resolved" nor
"integrity" field. That's the case for the yallist dependency from the
vega-embed project.

In that cases, we don't want to patch anything or throw an error, but
instead we'll leave the lockfile element as it is in the patched lockfile.

Adding a example-project having a bundled dependency without a
resolved field to test this behavior.