[v2][bugfix] Remove peerDependencies from produced lockfiles
As per
https://github.com/npm/npm/blob/2e3776bf5676bc24fec6239a3420f377fe98acde/doc/files/package.json.md#peerdependencies,
peerDependencies are only used to express compatibility. However, it
seems npm still does network access to fetch some information about
some of them, which fails in the sandbox.

This network access happen when some peer-dependencies are not
available through the node-modules spec, ie. their build receipe are
not available in the lockfile. We don't have enough informations to
fetch those: we lack both the resolved link and their integrity.

These dependencies do not influence the node_modules/ directory in any
ways. npmlock2nix only purpose is to build this directory, meaning we
can safely ignore the peer-dependencies during the node_modules/ nix
build.

This can be observed in @storybook/react, which contains
require-from-string@"^2.0.2" as peerDependency only. In the lockfile,
require-from-string isn't mentioned as a dependency, yet npm still
tries to look up information:

   npm WARN ERESOLVE overriding peer dependency0m Completed in 0ms
   npm WARN While resolving: @storybook/react@6.5.10
   npm WARN Found: require-from-string@undefined
   npm WARN node_modules/require-from-string
   npm WARN
   npm WARN Could not resolve dependency:
   npm WARN peer require-from-string@"^2.0.2" from @storybook/react@6.5.10
   npm WARN node_modules/@storybook/react
   npm WARN   dev @storybook/react@"^6.5.9" from the root project

Adding a unit test checking we indeed remove the peerDependencies from
the original lockfile.