[v2][bugfix] Strip out URL query strings in resolved fields
Npm gets confused when a lockfile file:// resolve field contains a
question mark in its name. It interprets that as a query string and
strips it away.

IE.

For a dependency like:

{
  "dep": {
    "version": "4.2.0",
    "resolved": "file:///nix/store/hash-blahblah.tgz?querystring=true",
    "integrity": "sha512-gHwtGSeteSl3LiSOk+rIENiVjI7yaMTYcxqroXZxErstz/5WcZV5Wme+8XCYBB7yLhMiWPvNlDS9Nr4urADIdQ==",
  },
}

Npm will try to open hash-blahblah.tgz instead of the actual file
hash-blahblah.tgz?querystring=true.

We go around this issue by making sure we remove anything that could
be interpreted as a query string by Npm in the store name.